+++ tests/auth_form_test.py
@@ -3,8 +3,12 @@ from http import HTTPStatus
 from urllib.parse import urlencode
 
 import flask.testing
+import httpx
 import pytest
 from faker import Faker
+from playwright.sync_api import Page, expect
+
+from oidc_provider_mock._client_lib import OidcClient
 
 from .conftest import use_provider_config
 
@@ -87,3 +91,40 @@ def test_missing_sub_parameter(client: flask.testing.FlaskClient):
     })
     response = client.post(f"/oauth2/authorize?{query}")
     assert "The field is missing" in response.text
+
+
+def test_authorized_users_buttons_appear_xxx(oidc_server: str, page: Page):
+    """Test that authorizing 3 users creates buttons on the auth page"""
+
+    client = OidcClient(
+        id=str(faker.uuid4()),
+        secret=faker.password(),
+        redirect_uri=faker.uri(schemes=["https"]),
+        issuer=oidc_server,
+    )
+
+    subject0 = faker.email()
+    subject1 = faker.email()
+
+    for sub in [subject0, subject1, subject0]:
+        httpx.post(
+            client.authorization_url(state=faker.password()),
+            data={"sub": sub},
+        )
+
+    page.goto(f"{oidc_server}/oidc/login")
+    page.get_by_role("button", name="Start").click()
+
+    expect(
+        page.get_by_role("heading", name="Authenticate previous users")
+    ).to_be_visible()
+
+    auth_buttons = page.get_by_role(
+        "form", name="Authenticate previous users"
+    ).get_by_role("button")
+
+    expect(auth_buttons.nth(0)).to_have_accessible_name(subject0)
+    expect(auth_buttons.nth(1)).to_have_accessible_name(subject1)
+
+    auth_buttons.nth(0).click()
+    expect(page.locator("body")).to_contain_text(f"You’re logged in as {subject0}")